+++ src/middleware.ts
@@ -1,0 +1,17 @@
+import createMiddleware from "next-intl/middleware";
+
+export default createMiddleware({
+  // A list of all locales that are supported
+  locales: ["en", "ko"],
+
+  // Used when no locale matches
+  defaultLocale: "en",
+
+  // Always prefix the pathname with locale
+  localePrefix: "always",
+});
+
+export const config = {
+  // Match only internationalized pathnames
+  matcher: ["/", "/(ko|en)/:path*"],
+};